requirements.txt
flask
werkzeug
pypdf
python-docx
nltk